Combine InstructorQuestionSettings and QuestionSettingsForm into a single component
