--- apps/prairielearn/assets/scripts/esm-bundles/hydrated-components/InstructorQuestionSettings.ts
@@ -1,5 +1,0 @@
-import { registerHydratedComponent } from '@prairielearn/react/hydrated-component';
-
-import { InstructorQuestionSettings } from '../../../../src/pages/instructorQuestionSettings/instructorQuestionSettings.html.js';
-
-registerHydratedComponent(InstructorQuestionSettings);
