+++ apps/prairielearn/assets/scripts/esm-bundles/hydrated-components/InstructorQuestionSettingsForm.ts
@@ -1,0 +1,5 @@
+import { registerHydratedComponent } from '@prairielearn/react/hydrated-component';
+
+import { InstructorQuestionSettingsForm } from '../../../../src/pages/instructorQuestionSettings/instructorQuestionSettings.html.js';
+
+registerHydratedComponent(InstructorQuestionSettingsForm);
